[refactor] : refactoring BloodPackCountControl
