+++ BloodStore/Assets/Scenes/Store.unity
@@ -5812,7 +5812,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 25.275, y: -15.645}
-  m_SizeDelta: {x: 54.94, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &618109123
 MonoBehaviour:
@@ -6460,7 +6460,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 10.91, y: -15.645}
-  m_SizeDelta: {x: 21.12, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &693935756
 MonoBehaviour:
@@ -6609,7 +6609,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 28.695, y: -15.645}
-  m_SizeDelta: {x: 57.18, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &701029262
 MonoBehaviour:
@@ -7520,7 +7520,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &758839428
 RectTransform:
   m_ObjectHideFlags: 0
@@ -8616,7 +8616,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 9.28, y: -15.645}
-  m_SizeDelta: {x: 21.29, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1014936900
 MonoBehaviour:
@@ -9221,7 +9221,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 18.14, y: -15.645}
-  m_SizeDelta: {x: 38.04, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1044630563
 MonoBehaviour:
@@ -11018,7 +11018,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 29.02, y: -15.645}
-  m_SizeDelta: {x: 59.33, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1232130354
 MonoBehaviour:
@@ -11970,15 +11970,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f9af86fc6ad0a8e4e97d3d4acd654cea, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  texts:
-  - {fileID: 739068328}
-  - {fileID: 279676228}
-  - {fileID: 2116498191}
-  - {fileID: 1172546156}
-  - {fileID: 758511269}
-  - {fileID: 272911649}
-  - {fileID: 1001046882}
-  - {fileID: 2067409204}
+  textList: []
 --- !u!222 &1329378993
 CanvasRenderer:
   m_ObjectHideFlags: 0
@@ -13069,7 +13061,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 9.28, y: -15.645}
-  m_SizeDelta: {x: 17.6, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1454688191
 MonoBehaviour:
@@ -16636,7 +16628,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 45.11, y: -15.645}
-  m_SizeDelta: {x: 87.19, y: 36.41}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1912877443
 MonoBehaviour:
@@ -17267,15 +17259,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f9af86fc6ad0a8e4e97d3d4acd654cea, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  texts:
-  - {fileID: 1452789672}
-  - {fileID: 2008921645}
-  - {fileID: 1893595525}
-  - {fileID: 916707385}
-  - {fileID: 20551519}
-  - {fileID: 1627269998}
-  - {fileID: 419734603}
-  - {fileID: 2119646028}
+  textList: []
 --- !u!114 &2018299559
 MonoBehaviour:
   m_ObjectHideFlags: 0
